+++ PsiphonTunnel.xcframework/ios-arm64/PsiphonTunnel.framework/Modules/module.modulemap
@@ -1,0 +1,6 @@
+framework module PsiphonTunnel {
+  umbrella header "PsiphonTunnel.h"
+
+  export *
+  module * { export * }
+}
