+++ PsiphonTunnel.xcframework/ios-arm64/PsiphonTunnel.framework/build-git-commit.txt
@@ -1,0 +1,1 @@
+f4315ffa6c7887a61aa0d7c1aa5515a7e6f916aa
