--- PsiphonTunnel.xcframework/ios-arm64_armv7/PsiphonTunnel.framework/Modules/module.modulemap
@@ -1,6 +1,0 @@
-framework module PsiphonTunnel {
-  umbrella header "PsiphonTunnel.h"
-
-  export *
-  module * { export * }
-}
