--- PsiphonTunnel.xcframework/ios-arm64_armv7/PsiphonTunnel.framework/build-git-commit.txt
@@ -1,1 +1,0 @@
-f344b6c58f9b641e2915528bead0de1d3ccac9f5
