+++ PsiphonTunnel.xcframework/ios-x86_64-simulator/PsiphonTunnel.framework/build-git-commit.txt
@@ -1,1 +1,1 @@
-f344b6c58f9b641e2915528bead0de1d3ccac9f5
+f4315ffa6c7887a61aa0d7c1aa5515a7e6f916aa
